Move GetTypesTest to a typesmanager package
