+++ flow-model-generator/modules/flow-model-generator-ls-extension/src/test/resources/testng.xml
@@ -40,12 +40,14 @@ under the License.
             <class name="io.ballerina.flowmodelgenerator.extension.ServiceGeneratorTest"/>
             <class name="io.ballerina.flowmodelgenerator.extension.VisibleVariableTypesTest"/>
             <class name="io.ballerina.flowmodelgenerator.extension.ConfigVariablesTest"/>
-            <class name="io.ballerina.flowmodelgenerator.extension.GetTypesTest"/>
             <class name="io.ballerina.flowmodelgenerator.extension.ModuleNodesTest"/>
             <class name="io.ballerina.flowmodelgenerator.extension.ConfigVariablesUpdateTest"/>
             <class name="io.ballerina.flowmodelgenerator.extension.DataMapManagerTypesTest"/>
             <class name="io.ballerina.flowmodelgenerator.extension.ErrorHandlerGeneratorTest"/>
             <class name="io.ballerina.flowmodelgenerator.extension.GetEnclosedFunctionDefTest"/>
+
+            <!--types manager-->
+            <class name="io.ballerina.flowmodelgenerator.extension.typesmanager.GetTypesTest"/>
         </classes>
     </test>
 </suite>
